add fine-grained scroll-wheel zoom
tileset_zoom becomes a double and gains zoom_in_fine/zoom_out_fine
(x/÷1.25), bound to the mouse wheel by default and registered in the
look-around, aiming and vehicle-shape UIs.

z/Z now snap to the nearest power of two instead of doubling/halving, so
the discrete zoom stays on clean steps after fine adjustments.

+++ src/overmap_ui.cpp
@@ -1858,7 +1858,7 @@ static std::vector<tripoint_abs_omt> get_overmap_path_to( const tripoint_abs_omt
     }
 }
 
-static int overmap_zoom_level = DEFAULT_TILESET_ZOOM;
+static double overmap_zoom_level = DEFAULT_TILESET_ZOOM;
 
 static bool try_travel_to_destination( avatar &player_character, const tripoint_abs_omt curs,
                                        const tripoint_abs_omt dest, const bool driving )
